refactoring server a bit to be more amenable to agent code gen, having working jpa model gen
