+++ server/generate.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+source "$HOME/.sdkman/bin/sdkman-init.sh"
+
+sdk env
+mvn clean generate-sources
